doc: fix mark action zero value in mlx5 guide
[ upstream commit b8ee0a16cb844027495968941a101f37964073cc ]

The zero value in flow MARK action is reported in Rx datapath
as tagged with zero FDIR ID. Once packet is marked in flow engine
it will be always reported as tagged. For metadata only the zero
value means there is "no metadata" in the packet and the metadata
flag is not set for the case.

Fixes: 3ceeed9f7855 ("doc: update flow mark action in mlx5 guide")

Signed-off-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>